Cleared life manager
- Removed life manager game over launch and camera trigger launch
logic. Placed instead on camera game object
