+++ Assets/detectLifeLoss.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3b7a4bd9fdcfb49e38ebb17adc7cbe4e
+timeCreated: 1487861375
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
