+++ Assets/setGameObjects.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c6b81a0ce9812478dab8e962eb0d8e5d
+timeCreated: 1487859677
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
